+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## `0.2.7` - 2019-08-19
+
+### Changed
+
+- Upgraded to GDK for Unity version `0.2.7`
 ### Breaking Changes
 
 - Moved `ChosenDeploymentAlphaLocatorFlow` and `SessionConnectionFlowInitializer` into the `Fps.Connection` namespace and made them `public`. [#225](https://github.com/spatialos/gdk-for-unity-fps-starter-project/pull/225)